\% Generated by roxygen2 (4.0.2): do not edit by hand
\name{affy.pivot.table}
\alias{affy.pivot.table}
\title{merge expression and calls together}
\usage{
affy.pivot.table(exp, calls)
}
\arguments{
\item{exp}{a \code{matrix}-like object of expression levels}

\item{calls}{a \code{matrix}-like object of detection calls (P/M/A)}
}
\value{
none.
}
\description{
merge expression and calls together
}
\author{
Mark Cowley
}

