\name{convert.gsea.res2gct}
\alias{convert.gsea.res2gct}
\title{Convert a res object to a gct object.}
\usage{convert.gsea.res2gct(res)
}
\description{Convert a res object to a gct object.}
\value{a gct object.}
\author{Mark Cowley, 2009-12-18}
\arguments{\item{res}{a res object. see import.gsea.res}
}

