\name{export.gsea.rpt}
\alias{export.gsea.rpt}
\title{Export a GSEA rpt file...}
\usage{export.gsea.rpt(rpt, f)
}
\description{Export a GSEA rpt file}
\value{none. creates a rpt file}
\author{Mark Cowley, 2009-12-11}
\arguments{\item{rpt}{a rpt object (ie a list). see import.gsea.rpt}
\item{f}{the filename}
}

