\name{gsea.convert.clm2cls}
\alias{gsea.convert.clm2cls}
\title{Convert a clm file into a cls file.}
\usage{
  gsea.convert.clm2cls(clm.file, cls.file = NULL)
}
\arguments{
  \item{clm.file}{the path to a clm file}

  \item{cls.file}{if NULL, the cls file will be in the same
  dir as the clm, and will have the suffix cls.}
}
\value{
  invisibly return a cls file name
}
\description{
  Convert a clm file into a cls file.
}
\author{
  Mark Cowley, 2009-12-21
}

