\name{gsea.fix.output.index.html}
\alias{gsea.fix.output.index.html}
\title{fix the index...}
\usage{gsea.fix.output.index.html(index.html, debug=FALSE)
}
\description{fix the index.html file}
\author{Mark Cowley
@return none. fixes html files.}
\seealso{\code{\link{gsea.fix.output}}}
\arguments{\item{index.html}{the path to the index.html report}
\item{debug}{logical: debugging mode?}
}

