\% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gsea.length}
\alias{gsea.length}
\title{How many genesets in a GSEA object?}
\usage{
gsea.length(x)
}
\arguments{
\item{x}{a GSEA object}
}
\value{
an integer indicating how many genesets are present
}
\description{
How many genesets in a GSEA object?
}
\author{
Mark Cowley, 2009-10-16
}

