\% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gsea.which.gmt}
\alias{gsea.which.gmt}
\title{What gmt file was used?}
\usage{
gsea.which.gmt(dir)
}
\arguments{
\item{dir}{the dir that contains index.html}
}
\value{
eg "c2_cgp"
}
\description{
What gmt file was used? This parses the gmx file name from within the rpt
file
}
\author{
Mark Cowley, 2009-04-07
}

