\name{import.gsea.clm}
\alias{import.gsea.clm}
\title{Import a clm file...}
\usage{import.gsea.clm(f)
}
\description{Import a clm file}
\details{Import a GenePattern clm file. CLM files are often used to annotate GCT files.}
\author{Mark Cowley, 2009-12-07}
\seealso{\code{\link{import.gsea.gct}} \code{\link{import.gsea.cls}} 
\url{http://www.broadinstitute.org/cancer/software/genepattern/tutorial/gp_fileformats.html#clm}}
\arguments{\item{f}{character: the path to a valid clm file.}
}

