\% Generated by roxygen2 (4.0.2): do not edit by hand
\name{import.gsea.gmx}
\alias{import.gsea.gmx}
\title{Import a gmx file that is commonly used by Broad's GSEA.}
\usage{
import.gsea.gmx(x)
}
\arguments{
\item{x}{either the path to a gmx file, the dir that contains index.html
  from a GSEA PreRanked run, or a URL to the Broad Institute's website (see
  import.gsea.rpt)}
}
\value{
a named list where each element is a vector of gene symbols the
  names are the geneset names
}
\description{
Either specify the full path to a gmx file, or from a GSEA PreRanked
run, provide the path to the top-level directory that contains the
index.html
}
\author{
Mark Cowley, 2012-12-05
}

