\% Generated by roxygen2 (4.0.2): do not edit by hand
\name{import.gsea}
\alias{import.gsea}
\title{Import the results from a GSEA or GseaPreRanked analysis.}
\usage{
import.gsea(x, edb = FALSE)
}
\arguments{
\item{x}{The path to the GSEA results to import.
(1) a GSEA zip file;
(2) a vector of multiple GSEA zip files;
(3) a single directory containing GSEA results. There should be
  an index.html file in this directory;
(4) a vector of multiple GSEA directories like in (3)
(5) a directory which contains >=1 GSEA directories within it.}

\item{edb}{logical: Import the edb file? See Details.}
}
\value{
A GSEA \code{list}, with the following elements:
  \item{tt}{the combined (pos & neg) GSEA top table}
  \item{leading.edge}{a named list of leading edge genes}
 \item{rnk}{The pre-ranked file (if one was used)}
\item{rpt}{The GSEA report, listing the parameter values used during the GSEA run}
\item{edb}{[optional] The edb object, as an \code{XML} tree. See \code{\link{import.gsea.edb}}}
}
\description{
Import the output from a single GSEA PreRanked run, including: the
toptable, the leading edge genes, the rnk file used, and the report which
summarises the analysis settings used, and optionally, the edb file.
}
\details{
A flexible function which takes a directory of GSEA results, or a GSEA zip
file, and imports
the `top table', the leading edge genes, the pre-ranked file (if one was
used) and the report file which describes the parameters used when the GSEA
was run.

The top table is a combination of the up/down tables that GSEA makes,
sorted by decreasing NES, with additional `LEADING.EDGE.SIZE' and
`DIRECTION' columns.

The leading.edge is a named list, where the names are the geneset names,
and each list element is a vector of gene symbols, sorted in decreasing
significance (i.e. first gene is the most important gene for this geneset).

The edb file contains the machine readable output which the GSEA Desktop
tool can read. You should import it if you plan on subsetting or merging
GSEA results which are destined to be viewed in the Leading Edge Viewer
tool. Otherwise it is quite large, so best avoided unless you really need
it!
}
\examples{
# not run
# import.gsea("./c2_all.Gsea.1252055214188")
# import.gsea("./c2_all.Gsea.1252055214188", edb=TRUE)
# import.gsea(c("./c1_all.Gsea.1252052322484", "./c2_all.Gsea.1252055214188", "./c3_all.Gsea.1252086970993"))
# import.gsea("/path/to/dir/containing/lots/of/GSEA/dirs")
}
\author{
Mark Cowley
}
\seealso{
\code{\link{import.gsea.topTable}}
  \code{\link{import.gsea.leadingedge}} \code{\link{import.gsea.rpt}}
  \code{\link{import.gsea.rnk}} \code{\link{import.gsea.edb}}
}
\keyword{IO}
\keyword{file}

