\name{is.gsea.dir}
\alias{is.gsea.dir}
\title{is the argument a valid GSEA directory?}
\usage{
  is.gsea.dir(x)
}
\arguments{
  \item{x}{path to a directory}
}
\value{
  \code{TRUE} if \code{x} is the top-level GSEA result
  directory, \code{FALSE} otherwise
}
\description{
  is the argument a valid GSEA directory?
}
\author{
  Mark Cowley, 2011-08-04
}

