\name{is.gsea.gct.file}
\alias{is.gsea.gct.file}
\title{is.gsea.gct.file}
\usage{
  is.gsea.gct.file(f)
}
\arguments{
  \item{f}{the path to a file}
}
\value{
  \code{TRUE} if \code{f} is a GCT file, \code{FALSE}
  otherwise
}
\description{
  is.gsea.gct.file
}
\author{
  Mark Cowley, 2011-10-21
}

