\name{is.gsea.list}
\alias{is.gsea.list}
\title{is x a gsea list, that is, a list containing gsea results?}
\usage{is.gsea.list(x)
}
\description{is \code{x} a gsea list, that is, a list containing gsea results?}
\author{Mark Cowley, 2008-05-13}
\value{logical: \code{TRUE} if x is a list, and contains only GSEA objects, \code{FALSE} otherwise.}
\arguments{\item{x}{an object}
}

