\name{list2df}
\alias{list2df}
\title{take a ragged list (ie a list of vectors of different lengths) and return a...}
\usage{list2df(x)
}
\description{take a ragged list (ie a list of vectors of different lengths) and return a
\code{data.frame}, filling the shorter vectors with \code{NA}'s}
\value{a \code{data.frame}, filled with \code{NA}'s where necessary}
\author{Mark Cowley, 25/2/08}
\arguments{\item{x}{a list of vectors}
}

