\name{plot.gsea.res}
\alias{plot.gsea.res}
\title{Boxplot the intensities for an array split by the detection call (P/M/A).}
\usage{plot.gsea.res(res, array=1)
}
\description{Boxplot the intensities for an array split by the detection call (P/M/A).}
\value{a boxplot}
\author{Mark Cowley, 2010-01-07}
\arguments{\item{res}{a gsea res object. see import.gsea.res}
\item{array}{which array to plot?}
}

