\% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_gsea.leadingedge.adjmat}
\alias{plot_gsea.leadingedge.adjmat}
\title{Generate a binary adjacency matrix image (like a heatmap with 2 colours).}
\usage{
plot_gsea.leadingedge.adjmat(x, main = "Leading Edge Adjacency Matrix",
  cluster = TRUE)
}
\arguments{
\item{x}{a GSEA object}

\item{main}{the title}

\item{cluster}{if TRUE then do a 2-way hierarchical clustering, else order
genes by decreasing frequency}
}
\value{
none. creates a plot.  Todo: introduce colour by looking up stat in
  the rnk object. handle the case of lots of genes nicer.
}
\description{
The occurence of genes in each geneset is represented by blue squares. Like
in the Broad's GSEA Leading Edge Viewer tool, genes are in columns, gensets
are in rows.
}
\author{
Mark Cowley, 2009-10-29
}

