\% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_gsea.leadingedge.HCL.1clusterPerPage}
\alias{plot_gsea.leadingedge.HCL.1clusterPerPage}
\title{Take a GSEA object, do an HCL between the leading edge genes, cut the HCL,
then plot 1 HCL per page of a PDF.}
\usage{
plot_gsea.leadingedge.HCL.1clusterPerPage(x,
  main = "Leading edge similarities", xlab = "",
  hclust.method = "complete", h = 0.99, ...)
}
\arguments{
\item{x}{a GSEA object}

\item{main}{the plot title}

\item{xlab}{the x label}

\item{hclust.method}{the hierarchical clustering method. see \code{\link{hclust}}}

\item{h}{numeric scalar or vector with heights where the tree should be
cut. (see cutree)}

\item{\dots}{see \code{\link{plot_gsea.leadingedge.HCL.simple}}}
}
\description{
Take a GSEA object, do an HCL between the leading edge genes, cut the HCL,
then plot 1 HCL per page of a PDF.
}
\author{
Mark Cowley, 2009-09-01
}

