\name{plot_gsea.res}
\alias{plot_gsea.res}
\title{Boxplot the intensities for an array split by the detection call (P/M/A).}
\usage{
  plot_gsea.res(res, array = 1)
}
\arguments{
  \item{res}{a gsea res object. see import.gsea.res}

  \item{array}{which array to plot?}
}
\value{
  a boxplot
}
\description{
  Boxplot the intensities for an array split by the
  detection call (P/M/A).
}
\author{
  Mark Cowley, 2010-01-07
}

