\name{subset.gsea}
\alias{subset.gsea}
\title{Subsetting GSEA objects.
Return subsets of GSEA objects which meet conditions, with consistent args to
\code{\link[base]{subset}}}
\usage{
  subset.gsea(x, subset, ...)
}
\arguments{
  \item{x}{object to be subsetted.}

  \item{subset}{logical expression indicating elements or
  rows to keep: missing values are taken as false.}

  \item{\dots}{unused}
}
\description{
  Subsetting GSEA objects. Return subsets of GSEA objects
  which meet conditions, with consistent args to
  \code{\link[base]{subset}}
}
\author{
  Mark Cowley, 2011-02-21
}

