\name{test.broad.ftp}
\alias{test.broad.ftp}
\title{Function to test whether a connection to BROAD's FTP/HTTP sites can be...}
\usage{test.broad.ftp(gsea.url)
}
\description{Function to test whether a connection to BROAD's FTP/HTTP sites can be
made.}
\value{TRUE/FALSE with an error message printed to stdout if no connection
can be made.}
\author{Mark Cowley, 2009-12-10}
\arguments{\item{gsea.url}{a URL to test}
}

