\name{export.gsea.edb}
\alias{export.gsea.edb}
\title{Export a GSEA edb object to an XML file.}
\usage{export.gsea.edb(edb, file)
}
\description{Export a GSEA edb object to an \code{XML} file.}
\value{creates an XML file}
\author{Mark Cowley, 2009-10-12}
\seealso{\code{\link{import.gsea.edb}}, \code{\link[XML]{saveXML}}}
\keyword{file}
\keyword{IO}
\arguments{\item{edb}{an edb object.}
\item{file}{the name of an xml file. Usually this is inside a dir called
edb, and ends in \dQuote{.edb}}
}

