\% Generated by roxygen2 (4.0.2): do not edit by hand
\name{export.gsea.gmt}
\alias{export.gsea.gmt}
\title{Export a gmt file
Genesets for BROAD's GSEA are usually represented either as \code{gmt} files, or
\code{gmx} files. These can be opened in \code{excel}, and the \code{gmt} contains the first row
being the geneset name, and subsequent rows containing genes within each
geneset. These genes do not have to be the same length in each geneset.}
\usage{
export.gsea.gmt(x, f, description = rep("na", length(x)))
}
\arguments{
\item{x}{A list of genesets. \code{x} must have names which are the geneset names,
and the elements are character vectors.}

\item{f}{the file name}

\item{description}{a character vector for the description column. it is recycled if its too short.
The order must match up with the rows of \code{x}.}
}
\value{
none. it creates a gmt file.
}
\description{
See the detailed file specification, and the \code{\link{import.gsea.gmt}}
function.
}
\author{
Mark Cowley, 2009-04-29
}
\references{
\url{http://www.broadinstitute.org/cancer/software/genepattern/tutorial/gp_fileformats.html#gmt}
}
\seealso{
\code{\link{import.gsea.gmt}}
}
\keyword{IO}
\keyword{file}

