\% Generated by roxygen2 (4.0.2): do not edit by hand
\name{export.gsea.rpt}
\alias{export.gsea.rpt}
\title{Export a GSEA rpt file}
\usage{
export.gsea.rpt(rpt, f)
}
\arguments{
\item{rpt}{a rpt object (ie a list). see import.gsea.rpt}

\item{f}{the filename}
}
\value{
none. creates a rpt file
}
\description{
Export a GSEA rpt file
}
\author{
Mark Cowley, 2009-12-11
}

