\name{gsea.fix.output.detailed.enrichment.results.html}
\alias{gsea.fix.output.detailed.enrichment.results.html}
\title{fix the "Detailed enrichment results in html format" 2 pages.}
\usage{
  gsea.fix.output.detailed.enrichment.results.html(pos.html,
  neg.html, debug = FALSE)
}
\arguments{
  \item{pos.html}{the path to the positive enrichemnt html
  report}

  \item{neg.html}{the path to the negative enrichemnt html
  report}

  \item{debug}{logical: debugging mode?}
}
\description{
  fix the "Detailed enrichment results in html format" 2
  pages.
}
\author{
  Mark Cowley @return none. fixes html files.
}
\seealso{
  \code{\link{gsea.fix.output}}
}

