\% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gsea.merge.edb}
\alias{gsea.merge.edb}
\title{Merge two GSEA edb objects.}
\usage{
gsea.merge.edb(edb1, edb2)
}
\arguments{
\item{edb1}{two edb XML objects. They should already have been renamed to
make the genesets unique. See gsea.rename.genesets}

\item{edb2}{two edb XML objects. They should already have been renamed to
make the genesets unique. See gsea.rename.genesets}
}
\value{
an edb XML object with as many elements as the two input edb's
  combined.
}
\description{
Merge two GSEA edb objects.
}
\author{
Mark Cowley, 2009-10-14
}

