\% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gsea.rename.genesets.edb}
\alias{gsea.rename.genesets.edb}
\title{Parse an edb XML object, and rename the genesets with optional prefix and
suffix.}
\usage{
gsea.rename.genesets.edb(edb, old.names, new.names)
}
\arguments{
\item{edb}{an XMLDocument. See \code{import.gsea.edb(..., edb=TRUE)}}

\item{old.names}{a vector of original geneset names. these should be found
within the edb names}

\item{new.names}{the new names, in the same order as the old names.}
}
\value{
An mutable \code{XMLInternalDOM} (ie an XML tree much like an XMLDocument
  but which can be modified) which is a copy of the original edb
  XMLDocument.
}
\description{
This is intended to be used by \code{gsea.rename.genesets}, which should calculate
\code{old.names} and \code{new.names}. Entries in \code{edb}, given that they are xml are in no
strict order (although they seem to be ordered from most -ve NES to most
+ve NES). Thus you need to supply the \code{old.names} is so that we can ensure
the entries in edb get renamed correctly.
}
\author{
Mark Cowley, 2009-10-06
}

