\% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gsea.reorder.gmt.by.rnk}
\alias{gsea.reorder.gmt.by.rnk}
\title{Reorder the genes within a gmt object to match the order of genes within a
rnk object}
\usage{
gsea.reorder.gmt.by.rnk(gmt, rnk)
}
\arguments{
\item{gmt}{a gmt object (ie a named list of genes)}

\item{rnk}{a rnk object (ie named list of gene scores, where genes are the
  names)}
}
\value{
a gmt just like input, but with the genes reordered to match the
  order inside the rnk
}
\description{
Reorder the genes within a gmt object to match the order of genes within a
rnk object
}
\author{
Mark Cowley, 2009-12-07
}

