\name{gsea.summarise}
\alias{gsea.summarise}
\title{Summarise a single GSEA result, or list of results in terms of how many...}
\usage{gsea.summarise(x)
}
\description{Summarise a single GSEA result, or list of results in terms of how many
genesets are changed.}
\value{a vector or data.frame of number of gene sets passing various
statistical thresholds See also: gsea.compare.runs.plot.sh}
\author{Mark Cowley, 2009-03-23}
\arguments{\item{x}{either a GSEA object, or a list of GSEA objects.}
}

