\name{gsea.summarise}
\alias{gsea.summarise}
\title{Summarise GSEA results}
\usage{
  gsea.summarise(x)
}
\arguments{
  \item{x}{either a GSEA object, or a list of GSEA
  objects.}
}
\value{
  a \code{vector} or \code{data.frame} of number of gene
  sets passing various statistical thresholds See also:
  \code{/path/to/metaGSEA/bin/gsea.compare.runs.plot.sh}
}
\description{
  Summarise a single GSEA result, or list of results in
  terms of how many genesets are changed.
}
\author{
  Mark Cowley, 2009-03-23
}

