\name{gsea.which.gmt}
\alias{gsea.which.gmt}
\title{What gmt file was used?}
\usage{gsea.which.gmt(dir)
}
\description{What gmt file was used? This parses the gmx file name from within the rpt
file}
\value{eg "c2_cgp"}
\author{Mark Cowley, 2009-04-07}
\arguments{\item{dir}{the dir that contains index.html}
}

