\% Generated by roxygen2 (4.0.2): do not edit by hand
\name{import.gsea.clm}
\alias{import.gsea.clm}
\title{Import a clm file}
\usage{
import.gsea.clm(f)
}
\arguments{
\item{f}{character: the path to a valid clm file.}
}
\value{
a \code{data.frame} with 3 columns: \sQuote{CEL}, \sQuote{Sample}, \sQuote{Class}
}
\description{
Import a GenePattern clm file. CLM files are often used to annotate GCT files.
}
\author{
Mark Cowley, 2009-12-07
}
\seealso{
\code{\link{import.gsea.gct}} \code{\link{import.gsea.cls}}
\url{http://www.broadinstitute.org/cancer/software/genepattern/tutorial/gp_fileformats.html#clm}
}

