\name{import.gsea.rnk}
\alias{import.gsea.rnk}
\title{Import a rnk file used for running GSEA PreRanked.}
\usage{
  import.gsea.rnk(x)
}
\arguments{
  \item{x}{the path to either a rnk file, or the dir that
  contains index.html}
}
\description{
  Either specify the full path to a rnk file, or from a
  GSEA PreRanked run, provide the path to the top-level
  directory that contains the index.html. NB, this imports
  the collapsed rnk file.
}
\author{
  Mark Cowley, 2009-04-28
}

