\% Generated by roxygen2 (4.0.2): do not edit by hand
\name{is.gsea.res.file}
\alias{is.gsea.res.file}
\title{is.gsea.res.file}
\usage{
is.gsea.res.file(f)
}
\arguments{
\item{f}{the path to a file}
}
\value{
\code{TRUE} if \code{f} is a RES file, \code{FALSE} otherwise
}
\description{
is.gsea.res.file
}
\author{
Mark Cowley, 2011-10-21
}

