\name{plot.gsea.leadingedge}
\alias{plot.gsea.leadingedge}
\title{Combination GSEA Leading Edge plot.}
\usage{
  plot.gsea.leadingedge(x, main)
}
\arguments{
  \item{x}{a GSEA list}

  \item{main}{the main plotting title}
}
\value{
  creates multiple plots.
}
\description{
  NB, it's only worth using on filtered GSEA data - see
  gsea.filter(<>, N=50, direction="up") See also
  plot.gsea.leadingedge.HCL, plot.gsea.leadingedge.heatmap,
  plot.gsea.leadingedge.barplot
}
\author{
  Mark Cowley, 2009-10-29
}

