\name{test.broad.ftp}
\alias{test.broad.ftp}
\title{test whether a connection to BROAD's FTP/HTTP sites can be
made.}
\usage{
  test.broad.ftp(gsea.url)
}
\arguments{
  \item{gsea.url}{a URL to test}
}
\value{
  logical(1) with an error message printed to stdout if no
  connection can be made.
}
\description{
  test whether a connection to BROAD's FTP/HTTP sites can
  be made.
}
\author{
  Mark Cowley, 2009-12-10
}

